Use correct formatting invocations for respective dispatchers
This commit addresses the crashing issues with SLF4J and other loggers during testing.
It uses the correct formatting invocations for every individual dispatcher, and only that invocation,
to prevent the final message being passed through to a different formatting invocation to cause a
failed log.

For example, given the following pseudo invocation of a facade implemented by a dispatcher:
    var object = ...;
    // object -> "Example[name = \"Whatever\", test = {023410}]"

    log("%s", object);

As is obvious, the object is substituted in and String.valueOf called only if the level is enabled.
However, the issue lies in that the toString method of that Example object returns a string containing
the following: "{023410}" - as that facade uses String.format convention, this would not usually cause
an issue, and the program would have previously worked before installing this dispatcher to switch
logging backends.

However, previously, MessageFormat would have also been called, as Strings#format was used. This causes
an issue as this string is not compatible with MessageFormat and seems to be referencing the 023410th
argument.

Signed-off-by: Mechite <github@mechite.com>