Support recursive call structure in Throwables, develop test suite
This commit removes the empty LoggerSpecification as this is to be implemented further down the line with more research and development of the functionality and interactivity with different dispatchers. More maturity is required for a robust test suite to be developed here.

It also introduces IterablesSpecification and ThrowablesSpecification, whose introduction then demonstrated failures in the ability for Throwables to handle recursive call structures which was caused by the assumption that Stream.iterate handled recursiveness and checked Object#equals - this was not researched adequately and has now been fixed - 1u1 will not be released until everything has thorougly been researched and previously used in production already.

Signed-off-by: Mechite <github@mechite.com>