Add jsr305 annotation implementation, Strings#shuffle*, traversal iterators, remove serialization
This commit makes the project ready for rudimentary publication. Majority of things are implemented to a usable degree after this commit. The next commit would be to introduce an SLF4J collection implementation for stellar.logging in order to allow for stellar.logging to be used as an API in dependencies without being the only supported underlying logging implementation (it would be an implementation that delegates to SLF4J, but maintain the overhead of still having its own thread which can easily be solved by using Logger#executor and implementing a virtual thread executor service).

Signed-off-by: Mechite <github@mechite.com>